chore(build): move build-bundle scripts to vite/ directory
